reorganization of types
