+++ src/components/lib/icons.types.ts
@@ -1,0 +1,5 @@
+export interface InlineIcon {
+  dangerouslySetInnerHTML: {
+    __html: string
+  }
+}
