+++ src/components/lib/styles.types.ts
@@ -1,0 +1,5 @@
+export enum TOOLBAR_COLOR {
+  NONE = 0,
+  HOVER = 1,
+  ACTIVE = 2
+}
