+++ src/lib/settings-storage.types.ts
@@ -1,0 +1,6 @@
+export type ViewportOption = "HD" | "FHD" | "QHD" | "UHD"
+
+export enum SETTINGS_KEY {
+  VIEWPORT = "VIEWPORT",
+  FONT_SIZE = "FONT_SIZE"
+}
