+++ src/lib/validation.ts
@@ -1,16 +1,5 @@
-import { TimedTrigger } from "./TriggerData";
-
-export interface ValueChange {
-  prev?: number | boolean
-  new: string | boolean
-}
-
-export interface Constraint {
-  DEFAULT: boolean | number
-  INT?: boolean
-  MIN?: number
-  MAX?: number
-}
+import { TimedTrigger } from "../lib/TriggerDataManager.types";
+import { ValueChange, Constraint } from "./validation.types";
 
 export const CONSTRAINTS = {
   INTERPOLATE: {
