+++ src/lib/validation.types.ts
@@ -1,0 +1,11 @@
+export interface ValueChange {
+  prev?: number | boolean
+  new: string | boolean
+}
+
+export interface Constraint {
+  DEFAULT: boolean | number
+  INT?: boolean
+  MIN?: number
+  MAX?: number
+}
